pyproject.toml
[project]
name = "breakdown"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "arviz<1.0.0",
    "dbt-metricflow>=0.10.1",
    "dbt-sl-sdk[sync]>=0.13.2",
    "fastapi>=0.135.2",
    "networkx>=3.6.1",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pydantic>=2.12.5",
    "pymc>=5.28.0",
    "pyyaml>=6.0.3",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
